third_party/WebKit/LayoutTests/external/wpt/html/webappapis/scripting/processing-model-2/body-onerror-compile-error-data-url-expected.txt

This is a testharness.js-based test.
FAIL <body onerror> - compile error in <script src=data:...> assert_equals: first arg expected "string" but got "object"
FAIL <body onerror> - compile error in <script src=data:...> (column) assert_equals: fourth arg expected "number" but got "undefined"
Harness: the test ran to completion.
